Added Blitz stance and associated skills.

+++ core/src/com/majalis/character/Technique.java
@@ -43,7 +43,7 @@ public class Technique {
 		// this is temporarily to prevent struggling from failing to work properly on the same term an eruption or knot happens
 		if (isSuccessful) isSuccessful = otherTechnique.getForceStance() == null || otherTechnique.getForceStance() == Stance.KNOTTED || otherTechnique.getForceStance() == Stance.KNEELING;
 		
-		return new Attack(isSuccessful, technique.getName(), (int)(getDamage() * blockMod), ((strength + technique.getPowerMod()) * technique.getKnockdown())/2, technique.isHealing() ? strength + technique.getPowerMod() : 0, technique.isTaunt() ? strength + technique.getPowerMod() : 0, technique.isGrapple() ? strength + technique.getPowerMod() : 0, technique.getClimaxType(), getForceStance());
+		return new Attack(isSuccessful, technique.getName(), (int)(getDamage() * blockMod), ((int) ((strength + technique.getPowerMod()) * technique.getKnockdown()))/2, technique.isHealing() ? strength + technique.getPowerMod() : 0, technique.isTaunt() ? strength + technique.getPowerMod() : 0, technique.isGrapple() ? strength + technique.getPowerMod() : 0, technique.getClimaxType(), getForceStance());
 	}
 	
 	private boolean isBlockable() {
